+++ tests.py
@@ -1,16 +1,12 @@
-import io
 import json
 import logging
 import os
-import sys
-import unittest
+from unittest import mock
 import uuid
 
-from tornado import httpclient, httputil, testing, web
-import mock
-import umsgpack
-
 from sprockets.mixins import http
+from tornado import httpclient, httputil, testing, web
+import umsgpack
 
 LOGGER = logging.getLogger(__name__)
 
@@ -25,7 +21,7 @@ def decode(value):
     if isinstance(value, list):
         return [decode(v) for v in value]
     elif isinstance(value, dict):
-        return dict([(decode(k), decode(v)) for k, v in value.items()])
+        return {decode(k): decode(v) for k, v in value.items()}
     elif isinstance(value, bytes):
         return value.decode('utf-8')
     return value
@@ -83,8 +79,13 @@ class TestHandler(web.RequestHandler):
                 'traceback': None})
         else:
             body = self.get_request_body()
-            if isinstance(body, dict) and 'response' in body:
-                return self.send_response(body['response'])
+            if isinstance(body, dict):
+                if 'link' in body:
+                    self.add_header('Link', body['link'])
+                if 'warning' in body:
+                    self.add_header('Warning', body['warning'])
+                if 'response' in body:
+                    return self.send_response(body['response'])
             self.send_response({'headers': dict(self.request.headers),
                                 'path': self.request.path,
                                 'args': self.request.arguments,
@@ -103,6 +104,9 @@ class TestHandler(web.RequestHandler):
         if content_type:
             LOGGER.debug('Setting response content-type: %r', content_type)
             self.set_header('Content-Type', content_type)
+        if 'Correlation-Id' in self.request.headers:
+            self.set_header(
+                'Correlation-Id', self.request.headers['Correlation-ID'])
         return self.write(decode(payload))
 
     def status_code(self):
@@ -124,6 +128,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
     def create_mixin(self, add_correlation=True):
         mixin = http.HTTPClientMixin()
         mixin.application = self._app
+        mixin.settings = self._app.settings
         mixin.request = httputil.HTTPServerRequest(
             'GET', 'http://test:9999/test',
             headers=httputil.HTTPHeaders(
@@ -131,7 +136,48 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
                 add_correlation else {}))
         return mixin
 
-    @testing.gen_test()
+    @testing.gen_test
+    def test_consumer_user_agent(self):
+
+        class Process:
+            def __init__(self):
+                self.consumer_version = '1.1.1'
+
+        class Consumer(http.HTTPClientMixin):
+            def __init__(self, *args, **kwargs):
+                super().__init__(*args, **kwargs)
+                self.name = 'consumer'
+                self.process = Process()
+
+        consumer = Consumer()
+        response = yield consumer.http_fetch(
+            self.get_url('/test?foo=bar&status_code=200'))
+        self.assertTrue(response.ok)
+        self.assertEqual(response.code, 200)
+        self.assertEqual(response.attempts, 1)
+        self.assertEqual(
+            response.body['headers'].get('User-Agent'), 'consumer/1.1.1')
+
+    @testing.gen_test
+    def test_consumer_user_agent_error(self):
+
+        class Consumer(http.HTTPClientMixin):
+            def __init__(self, *args, **kwargs):
+                super().__init__(*args, **kwargs)
+                self.name = 'consumer'
+                self.process = True
+
+        consumer = Consumer()
+        response = yield consumer.http_fetch(
+            self.get_url('/test?foo=bar&status_code=200'))
+        self.assertTrue(response.ok)
+        self.assertEqual(response.code, 200)
+        self.assertEqual(response.attempts, 1)
+        self.assertEqual(
+            response.body['headers'].get('User-Agent'),
+            'sprockets.mixins.http/{}'.format(http.__version__))
+
+    @testing.gen_test
     def test_default_user_agent(self):
         mixin = http.HTTPClientMixin()
         response = yield mixin.http_fetch(
@@ -143,7 +189,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
             response.body['headers'].get('User-Agent'),
             'sprockets.mixins.http/{}'.format(http.__version__))
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_default_user_agent_with_partial_config(self):
         del self._app.settings['version']
         response = yield self.mixin.http_fetch(
@@ -151,11 +197,13 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertTrue(response.ok)
         self.assertEqual(response.code, 200)
         self.assertEqual(response.attempts, 1)
+        self.assertGreater(response.duration, 0)
+        self.assertEqual(response.code, response.raw.code)
         self.assertEqual(
             response.body['headers'].get('User-Agent'),
             'sprockets.mixins.http/{}'.format(http.__version__))
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_socket_errors(self):
         with mock.patch(
                 'tornado.httpclient.AsyncHTTPClient.fetch') as fetch:
@@ -164,8 +212,13 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertFalse(response.ok)
         self.assertEqual(response.code, 599)
         self.assertEqual(response.attempts, 3)
+        self.assertIsNone(response.body)
+        self.assertIsNone(response.headers)
+        self.assertIsNone(response.links)
+        self.assertIsNone(response.raw)
+        self.assertTrue([isinstance(e, OSError) for e in response.exceptions])
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_without_correlation_id_behavior(self):
         mixin = self.create_mixin(False)
         response = yield mixin.http_fetch(
@@ -174,7 +227,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertEqual(response.code, 502)
         self.assertEqual(response.attempts, 3)
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_get(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test?foo=bar&status_code=200'))
@@ -187,8 +240,9 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
             response.body['headers'].get('User-Agent'), 'test/0.1.0')
         self.assertDictEqual(response.body['args'],
                              {'foo': ['bar'], 'status_code': ['200']})
+        self.assertIsNone(response.links)
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_post(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test'),
@@ -205,7 +259,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertDictEqual(response.body['body'],
                              {'foo': 'bar', 'status_code': 200})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_get_json(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test?foo=bar&status_code=200'),
@@ -221,7 +275,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertDictEqual(response.body['args'],
                              {'foo': ['bar'], 'status_code': ['200']})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_get_custom_user_agent(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test?foo=bar&status_code=200'),
@@ -238,7 +292,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertDictEqual(response.body['args'],
                              {'foo': ['bar'], 'status_code': ['200']})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_post_html(self):
         expectation = '<html>foo</html>'
         response = yield self.mixin.http_fetch(
@@ -256,7 +310,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
             response.body['headers'].get('User-Agent'), 'test/0.1.0')
         self.assertEqual(response.body['body'], expectation)
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_post_json(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test'),
@@ -275,7 +329,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertDictEqual(response.body['body'],
                              {'foo': 'bar', 'status_code': 200})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_post_custom_user_agent(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test'),
@@ -295,7 +349,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertDictEqual(response.body['body'],
                              {'foo': 'bar', 'status_code': 200})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_post_msgpack(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test'),
@@ -313,7 +367,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertDictEqual(response.body['body'],
                              {'foo': 'bar', 'status_code': 200})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_post_pre_serialized_json(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test'),
@@ -331,7 +385,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertDictEqual(response.body['body'],
                              {'foo': 'bar', 'status_code': 200})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_post_pre_serialized_msgpack(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test'),
@@ -348,16 +402,18 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
             response.body['headers'].get('User-Agent'), 'test/0.1.0')
         self.assertDictEqual(response.body['body'],
                              {'foo': 'bar', 'status_code': 200})
+        self.assertEqual([r.code for r in response.history], [200])
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_rate_limiting_behavior(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/error?status_code=429'))
         self.assertFalse(response.ok)
         self.assertEqual(response.code, 429)
         self.assertEqual(response.attempts, 3)
+        self.assertEqual([r.code for r in response.history], [429, 429, 429])
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_error_response(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/error?status_code=400&message=Test%20Error'))
@@ -366,16 +422,16 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         self.assertEqual(response.attempts, 1)
         self.assertEqual(response.body, 'Test Error')
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_error_retry(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/error?status_code=502'))
         self.assertFalse(response.ok)
         self.assertEqual(response.code, 502)
         self.assertEqual(response.attempts, 3)
-        self.assertEqual(response.body, b'')
+        self.assertIsNone(response.body)
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_unsupported_content_type(self):
         with self.assertRaises(ValueError):
             yield self.mixin.http_fetch(
@@ -384,20 +440,20 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
                 body=['foo', 'bar'],
                 request_headers={'Content-Type': 'text/html'})
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_unsupported_accept(self):
         expectation = '<html>foo</html>'
         response = yield self.mixin.http_fetch(
-                self.get_url('/test?content_type=text/html'),
-                method='POST',
-                body={'response': expectation},
-                request_headers={'Accept': 'text/html',
-                                 'Content-Type': 'application/json'})
+            self.get_url('/test?content_type=text/html'),
+            method='POST',
+            body={'response': expectation},
+            request_headers={'Accept': 'text/html',
+                             'Content-Type': 'application/json'})
         self.assertTrue(response.ok)
         self.assertEqual(response.headers['Content-Type'], 'text/html')
         self.assertEqual(response.body.decode('utf-8'), expectation)
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_allow_nonstardard_methods(self):
         response = yield self.mixin.http_fetch(
             self.get_url('/test'),
@@ -406,7 +462,7 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
             allow_nonstandard_methods=True)
         self.assertTrue(response.ok)
 
-    @testing.gen_test()
+    @testing.gen_test
     def test_max_clients_settings_supported(self):
         os.environ['HTTP_MAX_CLIENTS'] = '25'
         response = yield self.mixin.http_fetch(
@@ -416,18 +472,85 @@ class MixinTestCase(testing.AsyncHTTPTestCase):
         client = httpclient.AsyncHTTPClient()
         self.assertEqual(client.max_clients, 25)
 
-    @testing.gen_test()
-    @unittest.skipUnless(sys.version_info >= (3, ),
-                         'StringIO requires Python 3')
+    @testing.gen_test
     def test_missing_content_type(self):
         # Craft a response that lacks a Content-Type header.
-        request = httpclient.HTTPRequest(
+        body = 'Do not try to deserialize me'
+        response = yield self.mixin.http_fetch(
+            self.get_url('/test?foo=bar&status_code=200&content_type=0'),
+            'POST',
+            body=body,
+            content_type='text/plain')
+        self.assertTrue(response.ok)
+        del response._responses[-1].headers['Content-Type']
+        self.assertIsInstance(response.body, bytes)
+
+    @testing.gen_test
+    def test_get_link_header(self):
+        body = {'link': '<http://example.com/TheBook/chapter2>; '
+                'rel="previous"; '
+                'title="previous chapter"', 'status_code': 200}
+        response = yield self.mixin.http_fetch(
+            self.get_url('/test'),
+            method='POST',
+            body=body,
+            request_headers={'Accept': 'application/msgpack',
+                             'Content-Type': 'application/msgpack'})
+        self.assertTrue(response.ok)
+        self.assertEqual(response.code, 200)
+        self.assertEqual(response.body['headers'].get('Correlation-Id'),
+                         self.correlation_id)
+        self.assertEqual(response.attempts, 1)
+        self.assertEqual(
+            response.body['headers'].get('User-Agent'), 'test/0.1.0')
+        self.assertDictEqual(response.body['body'], body)
+
+        expectation = {
+            'target': 'http://example.com/TheBook/chapter2',
+            'rel': 'previous',
+            'title': 'previous chapter'
+        }
+        self.assertDictEqual(response.links[0], expectation)
+
+    @testing.gen_test
+    def test_get_warning_header(self):
+        body = {'warning': '110 anderson/1.3.37 "Response is stale"',
+                'status_code': 200}
+        with mock.patch.object(http.LOGGER, 'warning') as warning:
+            response = yield self.mixin.http_fetch(
+                self.get_url('/test'),
+                method='POST',
+                body=body,
+                request_headers={'Accept': 'application/msgpack',
+                                 'Content-Type': 'application/msgpack'})
+            warning.assert_called_once()
+
+        self.assertTrue(response.ok)
+        self.assertEqual(response.code, 200)
+        self.assertEqual(response.body['headers'].get('Correlation-Id'),
+                         self.correlation_id)
+        self.assertEqual(response.attempts, 1)
+        self.assertEqual(
+            response.body['headers'].get('User-Agent'), 'test/0.1.0')
+        self.assertDictEqual(response.body['body'], body)
+
+    @testing.gen_test
+    def test_correlation_id_attribute(self):
+        mixin = http.HTTPClientMixin()
+        mixin.correlation_id = str(uuid.uuid4())
+        response = yield mixin.http_fetch(
             self.get_url('/test?foo=bar&status_code=200'))
-        response = httpclient.HTTPResponse(
-            request, code=200, headers={},
-            buffer=io.StringIO('Do not try to deserialize me.'))
-        # Try to deserialize that response. It should not raise an exception.
-        try:
-            response_body = self.mixin._http_resp_deserialize(response)
-        except KeyError:
-            self.fail('http_fetch raised KeyError!')
+        self.assertTrue(response.ok)
+        self.assertEqual(response.code, 200)
+        self.assertEqual(response.attempts, 1)
+        self.assertEqual(
+            response.headers['Correlation-Id'], mixin.correlation_id)
+
+    @testing.gen_test
+    def test_dont_retry(self):
+        response = yield self.mixin.http_fetch(
+            self.get_url('/error?status_code=429'), dont_retry={429})
+        self.assertFalse(response.ok)
+        self.assertEqual(response.code, 429)
+        self.assertEqual(response.attempts, 1)
+        self.assertEqual([r.code for r in response.history], [429])